+++ petanque/json_shell/client.mli
@@ -17,6 +17,12 @@ module S (C : Chans) : sig
   val toc :
     TableOfContents.Params.t -> (TableOfContents.Response.t, string) result
 
+  val get_state_at_pos :
+    GetStateAtPos.Params.t -> (GetStateAtPos.Response.t, string) result
+
+  val get_root_state :
+    GetRootState.Params.t -> (GetRootState.Response.t, string) result
+
   val start : Start.Params.t -> (Start.Response.t, string) result
   val run : RunTac.Params.t -> (RunTac.Response.t, string) result
   val goals : Goals.Params.t -> (Goals.Response.t, string) result