+++ petanque/json_shell/interp_shell.ml
@@ -15,6 +15,10 @@ let do_handle ~fn ~token action =
     let open Coq.Compat.Result.O in
     let* doc = fn ~token ~uri |> of_pet_err in
     handler ~token ~doc
+  | Action.Pos { uri; point; handler } ->
+    let open Coq.Compat.Result.O in
+    let* doc = fn ~token ~uri |> of_pet_err in
+    handler ~token ~doc ~point
 
 (* Duplicate with lsp_core *)
 let feedback_to_message fb =